LPS-70229 Extracting media queries to avoid duplication
