+++ adaptive-media-image-api/bnd.bnd
@@ -8,6 +8,7 @@ Export-Package:\
 	com.liferay.adaptive.media.image.exception,\
 	com.liferay.adaptive.media.image.finder,\
 	com.liferay.adaptive.media.image.html,\
+	com.liferay.adaptive.media.image.mediaquery,\
 	com.liferay.adaptive.media.image.model,\
 	com.liferay.adaptive.media.image.optimizer,\
 	com.liferay.adaptive.media.image.processor,\
