+++ adaptive-media-image-api/src/main/java/com/liferay/adaptive/media/image/mediaquery/Condition.java
@@ -1,0 +1,38 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.adaptive.media.image.mediaquery;
+
+/**
+ * @author Alejandro Tardín
+ */
+public class Condition {
+
+	public Condition(String attribute, String value) {
+		_attribute = attribute;
+		_value = value;
+	}
+
+	public String getAttribute() {
+		return _attribute;
+	}
+
+	public String getValue() {
+		return _value;
+	}
+
+	private final String _attribute;
+	private final String _value;
+
+}
